coap: use broker IP address as destination
The underlying coap messages layer stores destination 'as is'
with the intent to match incoming responses to requests.
The matching process compares responses source address,
typically an IP address, with requests ._destination.

This fails to match responses sent from the broker when requests
are sent using a (local) fqdn broker address.